lite rr scenario validator as a CustomProcessValidator
